few CUDA-related fixed

+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/shape/Shape.java
@@ -2798,7 +2798,7 @@ public class Shape {
      * @return
      */
     public static DataBuffer shapeOf(DataBuffer buffer) {
-        int rank = buffer.getInt(0);
+        int rank = (int) buffer.getLong(0);
         return Nd4j.createBuffer(buffer, 1, rank);
     }
